src/requirements.txt
Flask==2.1.1
joblib==1.1.0
kafka_python==2.0.2
numpy==1.21.4
requests==2.25.1
scikit_learn==1.0.2
